moving karpenter specific type exports back to index, from spi
